Left Hand Glove Final Render and Model Complete
This is the final model of the left-hand glove, which will point and click on the buttons to switch between tanks. I have essentially used the same processes for creating the right-hand glove, but built this left-hand model from scratch, since the pose and function of this glove required different structures and articulations.